server/requirements.txt
ecdsa
flask
flask-cors
flask-migrate
flask-sqlalchemy
gunicorn
lnd-grpc-client
lnurl
psycopg2
pyjwt
pyqrcode
python-dateutil
requests